:sparkles: added a floating action button to the experiment list
