+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,6 @@
     <string name="confirm_password_mandatory_field">Confirm password*</string>
     <string name="confirm_password">Confirm Password</string>
     <string name="add_experiment">Add experiment</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
